Move Lang trait in its own module.
