+++ src/terminal.rs
@@ -14,7 +14,7 @@ use room::*;
 use base::*;
 use base::Watchable;
 use fight::DamageRes;
-use lang::lang;
+use lang::t;
 
 pub type CommandFn = Box<Fn(&mut Game) -> result::Result<bool, Box<Error>>>;
 
@@ -63,7 +63,7 @@ impl Terminal {
         if let Some(command) = self.commands.get(keyword) {
             command(&mut self.game)
         } else {
-            Err(Box::new(GameError::GeneralError(lang().command_not_found(keyword))))
+            Err(Box::new(GameError::GeneralError(t().command_not_found(keyword))))
         }
     }
     pub fn run(&mut self) {
@@ -86,34 +86,34 @@ pub fn cmd_look() -> CommandFn {
 }
 pub fn cmd_look_item() -> CommandFn {
     Box::new(|game| {
-        let keyword = input_string(&lang().item_prompt())?;
+        let keyword = input_string(&t().item_prompt())?;
         let room: &Room = game.player_room();
         let item = room.get_item(&keyword)
-            .ok_or(lang().keyword_not_found(&keyword))?;
+            .ok_or(t().keyword_not_found(&keyword))?;
         println!("{}", item.watch());
         Ok(false)
     })
 }
 pub fn cmd_look_actor() -> CommandFn {
     Box::new(|game| {
-        let keyword = input_string(&lang().actor_prompt())?;
+        let keyword = input_string(&t().actor_prompt())?;
         let room: &Room = game.player_room();
         let actor = room.get_actor(&keyword)
-            .ok_or(lang().keyword_not_found(&keyword))?;
+            .ok_or(t().keyword_not_found(&keyword))?;
         println!("{}", actor.watch());
         Ok(false)
     })
 }
 pub fn cmd_room_ref() -> CommandFn {
     Box::new(|game| {
-        println!("{}", lang().room_key_response(&game.room_ref));
+        println!("{}", t().room_key_response(&game.room_ref));
         Ok(false)
     })
 }
 pub fn cmd_add_exit() -> CommandFn {
     Box::new(|mut game| {
-        let exit_label = input_string(&lang().exit_label_prompt())?;
-        let room_id = input_string(&lang().room_key_prompt())?;
+        let exit_label = input_string(&t().exit_label_prompt())?;
+        let room_id = input_string(&t().room_key_prompt())?;
         let room_id: usize = room_id.parse()?;
         game.player_room_mut().add_exit(Exit {
             label: exit_label,
@@ -124,18 +124,18 @@ pub fn cmd_add_exit() -> CommandFn {
 }
 pub fn cmd_add_room() -> CommandFn {
     Box::new(|mut game| {
-        let title = input_string(&lang().room_title_prompt())?;
+        let title = input_string(&t().room_title_prompt())?;
         let key = game.rooms.add(Room::with_title(title));
-        println!("{}", lang().room_key_response(&key));
+        println!("{}", t().room_key_response(&key));
         Ok(false)
     })
 }
 pub fn cmd_move_player() -> CommandFn {
     Box::new(|mut game| {
-        let direction = input_string(&lang().exit_name_prompt())?;
+        let direction = input_string(&t().exit_name_prompt())?;
         let room_key = {
             let exit = game.player_room().get_exit(direction)
-                .ok_or(lang().exit_not_found_response())?;
+                .ok_or(t().exit_not_found_response())?;
             let dest_room_key = exit.room_key;
             dest_room_key
         };
@@ -148,7 +148,7 @@ pub fn cmd_move_player() -> CommandFn {
 }
 pub fn cmd_take() -> CommandFn {
     Box::new(|mut game| {
-        let item_key = input_string(&lang().item_prompt())?;
+        let item_key = input_string(&t().item_prompt())?;
         let player_ref = game.player_ref.clone();
         game.player_room_mut().actor_take(
             &player_ref,
@@ -159,7 +159,7 @@ pub fn cmd_take() -> CommandFn {
 }
 pub fn cmd_drop() -> CommandFn {
     Box::new(|mut game| {
-        let item_key = input_string(&lang().item_prompt())?;
+        let item_key = input_string(&t().item_prompt())?;
         let player_ref = game.player_ref.clone();
         game.player_room_mut().actor_drop(
             &player_ref,
@@ -171,16 +171,16 @@ pub fn cmd_drop() -> CommandFn {
 
 pub fn cmd_attack() -> CommandFn {
     Box::new(|mut game| {
-        let actor_key = input_string(&lang().actor_prompt())?;
+        let actor_key = input_string(&t().actor_prompt())?;
         let player_ref = game.player_ref.clone();
         let mut room = game.player_room_mut();
         match room.attack(&player_ref, &actor_key)? {
             DamageRes::NoDamage =>
-                println!("{}", lang().no_damage_response()),
+                println!("{}", t().no_damage_response()),
             DamageRes::Default(damage) =>
-                println!("{}", lang().damage_response(damage)),
+                println!("{}", t().damage_response(damage)),
             DamageRes::Dead =>
-                println!("{}", lang().dead_response())
+                println!("{}", t().dead_response())
         };
         Ok(false)
     })
@@ -188,11 +188,11 @@ pub fn cmd_attack() -> CommandFn {
 
 pub fn cmd_edit_room() -> CommandFn {
     Box::new(|mut game| {
-        let title = input_string(&lang().room_title_prompt())?;
-        println!("{}", lang().multiline_info());
+        let title = input_string(&t().room_title_prompt())?;
+        println!("{}", t().multiline_info());
         let description = read_multiline(
-            &lang().description_prompt(),
-            &lang().default_multiline_term())?;
+            &t().description_prompt(),
+            &t().default_multiline_term())?;
         let mut room = game.player_room_mut();
         room.title = title;
         room.description = description;
@@ -202,15 +202,15 @@ pub fn cmd_edit_room() -> CommandFn {
 
 pub fn cmd_save() -> CommandFn {
     Box::new(|game| {
-        println!("{}", lang().game_name_info());
-        let mut name = input_string(&lang().game_name_prompt())?;
+        println!("{}", t().game_name_info());
+        let mut name = input_string(&t().game_name_prompt())?;
         let failed_validation = name.chars().find(|c|
             !(*c >= 'a' && *c <= 'z' ||
                 *c >= 'A' && *c <= 'Z' ||
                 *c >= '0' && *c <= '9'));
         name.push_str(".json");
         if failed_validation.is_some() {
-            Err(Box::new(GameError::GeneralError(lang().invalid_characters_response())))
+            Err(Box::new(GameError::GeneralError(t().invalid_characters_response())))
         } else {
             let string = serde_json::to_string(game)?;
             let mut out = File::create(name)?;
@@ -221,15 +221,15 @@ pub fn cmd_save() -> CommandFn {
 }
 pub fn cmd_load() -> CommandFn {
     Box::new(|game| {
-        println!("{}", lang().game_name_info());
-        let mut name = input_string(&lang().game_name_prompt())?;
+        println!("{}", t().game_name_info());
+        let mut name = input_string(&t().game_name_prompt())?;
         let failed_validation = name.chars().find(|c|
             !(*c >= 'a' && *c <= 'z' ||
                 *c >= 'A' && *c <= 'Z' ||
                 *c >= '0' && *c <= '9'));
         name.push_str(".json");
         if failed_validation.is_some() {
-            Err(Box::new(GameError::GeneralError(lang().invalid_characters_response())))
+            Err(Box::new(GameError::GeneralError(t().invalid_characters_response())))
         } else {
             let file = File::open(name)?;
             let mut load_game = serde_json::from_reader(file)?;
